+++ NEWS.md
@@ -6,6 +6,7 @@ Update UBR and take Windows 11 26H1 into account.
 
 New presets/rules:
 
+* 2026/03/10 - `DisableAnonymousShareEnumeration`/`EnableAnonymousShareEnumeration`/`ViewAnonymousShareEnumeration` -> Disable Anonymous enumeration of shares
 * 2026/03/10 - `DisableAnonymousNameTranslation`/`EnableAnonymousNameTranslation`/`ViewAnonymousNameTranslation` -> Disable Anonymous SID/Name translation
 * 2026/03/04 - `DisableWinRMBasicAuth`/`EnableWinRMBasicAuth`/`ViewWinRMBasicAuth` -> Disable Windows Remote Management (WinRM) Basic Authentication (Client and Service)